Refactor to apply common code
 - Use the existing EntryPointStyle and ImagePullPolicy from the deployer
