--- src/test/java/org/springframework/cloud/deployer/scheduler/spi/kubernetes/ImagePullPolicyTests.java
@@ -1,45 +1,0 @@
-/*
- * Copyright 2018-2019 the original author or authors.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.springframework.cloud.deployer.scheduler.spi.kubernetes;
-
-import org.junit.Test;
-
-import static org.junit.Assert.assertEquals;
-
-/**
- * Tests for {@link ImagePullPolicy}.
- *
- * @author Chris Schaefer
- */
-public class ImagePullPolicyTests {
-	@Test
-	public void testInvalidImagePullPolicyDefaulting() {
-		ImagePullPolicy imagePullPolicy = ImagePullPolicy.relaxedValueOf("unknown");
-		assertEquals(ImagePullPolicy.IfNotPresent, imagePullPolicy);
-	}
-
-	@Test
-	public void testMatchingImagePullPolicy() {
-		ImagePullPolicy imagePullPolicy = ImagePullPolicy.relaxedValueOf("Always");
-		assertEquals(ImagePullPolicy.Always, imagePullPolicy);
-	}
-
-	@Test
-	public void testMixedCaseImagePullPolicy() {
-		ImagePullPolicy imagePullPolicy = ImagePullPolicy.relaxedValueOf("nEvEr");
-		assertEquals(ImagePullPolicy.Never, imagePullPolicy);
-	}
-}
